Created a function for read data entered by the user
The function takes one argument - the name of the coefficient. The function then allows the user to enter the value of the coefficient in the console. If the value is not valid, the function will repeat the reading until the user enters a valid number. The function also handles the case when a = 0.