Decomposed the function of finding roots
In the previous version of the function for solving quadratic equations, I solved the quadratic equation and immediately derived the result of the roots. This is not a good practice, because in the future we may want to use this feature to find roots, but not output them. To do this, I changed the function for solving square roots and now it solves the equation and returns an array with the results. And to display these results, I created a function "displayRoots", that takes the coefficients and the array of roots and displays them in the console.