The function of start file mode, the main file and the test file are created
The file mode function takes the path to a file with spaces separated by spaces and a new line at the end, and tries to read it. If the file format and path to it are correct - the function starts calculating the roots and displaying them on the screen. If the entered path to the file is incorrect - the function displays error messages.