Created the function of the interactive mode of the program
The function does not accept any parameters. The interactive mode function alternately asks the user for the values of the coefficients by calling the "getCoefficient" function, and then after entering the coefficients starts the function of finding the roots "solveEquation" and outputting the result to the console by calling function "displayRoots".